Legacy browser support (Microsoft Edge)
I have run into a problem working with windows dot-net web container, which is Microsoft Edge 12 based. `globalThis` is not supported in this browser and updating a dependency introduced latest `get-browser-rtc` into our project, making it fail to connect.

This fix should run a short "fallback cycle" from globalThis to window to global (for nodejs) and then fail if none exists.